requirements.txt
python-docx
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
groq
redis
fastapi
uvicorn
pydantic
